modify project structure
