+++ app/src/main/java/com/example/ebn/activities/LoginActivity.java
@@ -1,9 +1,8 @@
-package com.example.ebn;
+package com.example.ebn.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -14,20 +13,22 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.ebn.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
-public class Register extends AppCompatActivity {
+public class LoginActivity extends AppCompatActivity {
 
     EditText editTextEmail, editTextPassword;
-    Button buttonReg;
-    TextView loginNow;
+    Button buttonLog;
+    TextView registerNow;
     ProgressBar progressBar;
     FirebaseAuth mAuth;
 
+    @Override
     public void onStart() {
         super.onStart();
         // Check if user is signed in (non-null) and update UI accordingly.
@@ -42,25 +43,25 @@ public class Register extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
+        setContentView(R.layout.activity_login);
 
-        editTextEmail = findViewById(R.id.editTextEmailReg);
-        editTextPassword = findViewById(R.id.editTextPasswordReg);
-        buttonReg = findViewById(R.id.buttonReg);
-        loginNow = findViewById(R.id.loginNow);
-        progressBar = findViewById(R.id.progressBarReg);
+        editTextEmail = findViewById(R.id.editTextEmailLog);
+        editTextPassword = findViewById(R.id.editTextPasswordLog);
+        buttonLog = findViewById(R.id.buttonLog);
+        registerNow = findViewById(R.id.registerNow);
+        progressBar = findViewById(R.id.progressBarLog);
         mAuth = FirebaseAuth.getInstance();
 
-        loginNow.setOnClickListener(new View.OnClickListener() {
+        registerNow.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Login.class);
+                Intent intent = new Intent(getApplicationContext(), RegisterActivity.class);
                 startActivity(intent);
                 finish();
             }
         });
 
-        buttonReg.setOnClickListener(new View.OnClickListener() {
+        buttonLog.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 progressBar.setVisibility(View.VISIBLE);
@@ -69,29 +70,29 @@ public class Register extends AppCompatActivity {
                 password = String.valueOf(editTextPassword.getText());
 
                 if (TextUtils.isEmpty(email)) {         //check input empty or not
-                    Toast.makeText(Register.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
+                    Toast.makeText(LoginActivity.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
                     return;
                 }
 
                 if (TextUtils.isEmpty(email)) {         //check input empty or not
-                    Toast.makeText(Register.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
+                    Toast.makeText(LoginActivity.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
                     return;
                 }
 
-                mAuth.createUserWithEmailAndPassword(email, password)
+                mAuth.signInWithEmailAndPassword(email, password)
                         .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
                             @Override
                             public void onComplete(@NonNull Task<AuthResult> task) {
                                 progressBar.setVisibility(View.GONE);
                                 if (task.isSuccessful()) {
-                                    Toast.makeText(Register.this, "Account created.",
+                                    Toast.makeText(LoginActivity.this, "Login successful.",
                                             Toast.LENGTH_SHORT).show();
-                                    Intent intent = new Intent(getApplicationContext(),Login.class);
+                                    Intent intent = new Intent(getApplicationContext(),MainActivity.class);
                                     startActivity(intent);
                                     finish();
                                 } else {
                                     // If sign in fails, display a message to the user.
-                                    Toast.makeText(Register.this, "Authentication failed.",
+                                    Toast.makeText(LoginActivity.this, "Authentication failed.",
                                             Toast.LENGTH_SHORT).show();
                                 }
                             }
