+++ app/src/main/java/com/example/ebn/activities/RegisterActivity.java
@@ -1,9 +1,8 @@
-package com.example.ebn;
+package com.example.ebn.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -14,21 +13,21 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.ebn.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
-public class Login extends AppCompatActivity {
+public class RegisterActivity extends AppCompatActivity {
 
     EditText editTextEmail, editTextPassword;
-    Button buttonLog;
-    TextView registerNow;
+    Button buttonReg;
+    TextView loginNow;
     ProgressBar progressBar;
     FirebaseAuth mAuth;
 
-    @Override
     public void onStart() {
         super.onStart();
         // Check if user is signed in (non-null) and update UI accordingly.
@@ -43,25 +42,25 @@ public class Login extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        setContentView(R.layout.activity_register);
 
-        editTextEmail = findViewById(R.id.editTextEmailLog);
-        editTextPassword = findViewById(R.id.editTextPasswordLog);
-        buttonLog = findViewById(R.id.buttonLog);
-        registerNow = findViewById(R.id.registerNow);
-        progressBar = findViewById(R.id.progressBarLog);
+        editTextEmail = findViewById(R.id.editTextEmailReg);
+        editTextPassword = findViewById(R.id.editTextPasswordReg);
+        buttonReg = findViewById(R.id.buttonReg);
+        loginNow = findViewById(R.id.loginNow);
+        progressBar = findViewById(R.id.progressBarReg);
         mAuth = FirebaseAuth.getInstance();
 
-        registerNow.setOnClickListener(new View.OnClickListener() {
+        loginNow.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Register.class);
+                Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
                 startActivity(intent);
                 finish();
             }
         });
 
-        buttonLog.setOnClickListener(new View.OnClickListener() {
+        buttonReg.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 progressBar.setVisibility(View.VISIBLE);
@@ -70,29 +69,29 @@ public class Login extends AppCompatActivity {
                 password = String.valueOf(editTextPassword.getText());
 
                 if (TextUtils.isEmpty(email)) {         //check input empty or not
-                    Toast.makeText(Login.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
+                    Toast.makeText(RegisterActivity.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
                     return;
                 }
 
                 if (TextUtils.isEmpty(email)) {         //check input empty or not
-                    Toast.makeText(Login.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
+                    Toast.makeText(RegisterActivity.this, "Please enter your email!",Toast.LENGTH_SHORT).show();
                     return;
                 }
 
-                mAuth.signInWithEmailAndPassword(email, password)
+                mAuth.createUserWithEmailAndPassword(email, password)
                         .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
                             @Override
                             public void onComplete(@NonNull Task<AuthResult> task) {
                                 progressBar.setVisibility(View.GONE);
                                 if (task.isSuccessful()) {
-                                    Toast.makeText(Login.this, "Login successful.",
+                                    Toast.makeText(RegisterActivity.this, "Account created.",
                                             Toast.LENGTH_SHORT).show();
-                                    Intent intent = new Intent(getApplicationContext(),MainActivity.class);
+                                    Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
                                     startActivity(intent);
                                     finish();
                                 } else {
                                     // If sign in fails, display a message to the user.
-                                    Toast.makeText(Login.this, "Authentication failed.",
+                                    Toast.makeText(RegisterActivity.this, "Authentication failed.",
                                             Toast.LENGTH_SHORT).show();
                                 }
                             }
