+++ app/src/main/java/com/example/ebn/fragments/HomepageFragment.java
@@ -1,4 +1,4 @@
-package com.example.ebn;
+package com.example.ebn.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.ebn.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link RoomFragment#newInstance} factory method to
+ * Use the {@link HomepageFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class RoomFragment extends Fragment {
+public class HomepageFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class RoomFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public RoomFragment() {
+    public HomepageFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class RoomFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment RoomFragment.
+     * @return A new instance of fragment HomepageFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static RoomFragment newInstance(String param1, String param2) {
-        RoomFragment fragment = new RoomFragment();
+    public static HomepageFragment newInstance(String param1, String param2) {
+        HomepageFragment fragment = new HomepageFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class RoomFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_room, container, false);
+        return inflater.inflate(R.layout.fragment_homepage, container, false);
     }
 }
