+++ app/src/main/res/layout/fragment_account.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AccountFragment"
+    tools:context=".fragments.AccountFragment"
     android:background="@color/white">
 
     <!-- TODO: Update blank fragment layout -->
