+++ app/src/main/res/layout/fragment_homepage.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomepageFragment"
+    tools:context=".fragments.HomepageFragment"
     android:background="@color/teal_200">
 
     <!-- TODO: Update blank fragment layout -->
