+++ app/src/main/res/layout/fragment_qr_scan.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QrScanFragment"
+    tools:context=".fragments.QrScanFragment"
     android:background="@color/teal_200">
 
     <!-- TODO: Update blank fragment layout -->
